增加传递screenshot及属性，属性是否正确传递待验证

+++ MLN-iOS/MLNDevTool/Classes/Protobuf/Pbpoint.pbobjc.h
@@ -53,9 +53,9 @@ typedef GPB_ENUM(pbpoint_FieldNumber) {
 
 @interface pbpoint : GPBMessage
 
-@property(nonatomic, readwrite) float x;
+@property(nonatomic, readwrite) double x;
 
-@property(nonatomic, readwrite) float y;
+@property(nonatomic, readwrite) double y;
 
 @end
